feat(TRPC): add error handling function
