+++ apps/web/src/app/dashboard/files/UpdateDialog.tsx
@@ -13,7 +13,7 @@ import { Switch } from "@paperplane/ui/switch";
 import { useToast } from "@paperplane/ui/use-toast";
 import { Checkbox } from "@paperplane/ui/checkbox";
 import { api } from "#trpc/server";
-import { getTRPCError } from "@paperplane/utils";
+import { HandleTRPCFormError } from "#trpc/shared";
 
 export interface UpdateDialogProps {
 	/** The name (id) of the url */
@@ -54,16 +54,7 @@ export const UpdateDialog: React.FC<UpdateDialogProps> = ({ name, visible, passw
 			toast({ title: "File updated", description: "The file has been updated." });
 			form.reset(undefined, { keepDirty: false });
 		} catch (err) {
-			const parsedError = getTRPCError(err.message);
-			if (!parsedError) {
-				console.error(err);
-				form.setError("name", { message: "Unknown error, please try again later." });
-				return;
-			}
-
-			const inputField = parsedError.field as keyof z.infer<typeof FormSchema>;
-			if (Boolean(form.getValues()[inputField])) form.setError(inputField, { message: parsedError.message });
-			console.error(parsedError);
+			HandleTRPCFormError(err, form, "name");
 		}
 	}
